Fixed save to service rather than search
